+++ packages/components-js/projects/partials/tests/unit/specs/getIconLinks.spec.ts
@@ -1,5 +1,5 @@
 import { getIconLinks } from '../../../src';
-import type { IconNameCamelCase } from '@porsche-design-system/icons';
+import type { IconName } from '@porsche-design-system/icons';
 import { ICON_NAMES } from '@porsche-design-system/icons';
 import { paramCase } from 'change-case';
 import { render } from '@testing-library/react';
@@ -28,7 +28,7 @@ describe('format: html', () => {
   it('should return default link', () => {
     const result = getIconLinks();
     const regex = new RegExp(
-      `^<link rel=prefetch href=${baseHrefCom}/arrow-head-right.min.${hash}.svg as=image type=image/svg\\+xml crossorigin>$`
+      `^<link rel=prefetch href=${baseHrefCom}/arrow-right.min.${hash}.svg as=image type=image/svg\\+xml crossorigin>$`
     );
     expect(result).toMatch(regex);
   });
@@ -36,13 +36,13 @@ describe('format: html', () => {
   it('should return default link for china cdn', () => {
     const result = getIconLinks({ cdn: 'cn' });
     const regex = new RegExp(
-      `^<link rel=prefetch href=${baseHrefCn}/arrow-head-right.min.${hash}.svg as=image type=image/svg\\+xml crossorigin>$`
+      `^<link rel=prefetch href=${baseHrefCn}/arrow-right.min.${hash}.svg as=image type=image/svg\\+xml crossorigin>$`
     );
     expect(result).toMatch(regex);
   });
 
   it('should return multiple links', () => {
-    const result = getIconLinks({ icons: ['truck', 'volumeUp', 'mobile'] });
+    const result = getIconLinks({ icons: ['truck', 'volume-up', 'mobile'] });
     const regex = new RegExp(
       `^<link rel=prefetch href=${baseHrefCom}/truck.min.${hash}.svg as=image type=image/svg\\+xml crossorigin><link rel=prefetch href=${baseHrefCom}/volume-up.min.${hash}.svg as=image type=image/svg\\+xml crossorigin><link rel=prefetch href=${baseHrefCom}/mobile.min.${hash}.svg as=image type=image/svg\\+xml crossorigin>$`
     );
@@ -50,7 +50,7 @@ describe('format: html', () => {
     expect(result).toMatch(regex);
   });
 
-  ICON_NAMES.forEach((iconName: IconNameCamelCase) => {
+  ICON_NAMES.forEach((iconName: IconName) => {
     it(`should match regex for ['${iconName}']`, () => {
       const result = getIconLinks({ icons: [iconName] });
       const regex = new RegExp(
@@ -67,7 +67,7 @@ describe('format: jsx', () => {
   it('should return default link', () => {
     const { container } = render(getIconLinks({ format: 'jsx' }));
     const regex = new RegExp(
-      `^<link rel="prefetch" href="${baseHrefCom}/arrow-head-right.min.${hash}.svg" as="image" type="image/svg\\+xml" crossorigin="true">$`
+      `^<link rel="prefetch" href="${baseHrefCom}/arrow-right.min.${hash}.svg" as="image" type="image/svg\\+xml" crossorigin="true">$`
     );
     expect(container.innerHTML).toMatch(regex);
   });
@@ -75,20 +75,20 @@ describe('format: jsx', () => {
   it('should return default link for china cdn', () => {
     const { container } = render(getIconLinks({ format: 'jsx', cdn: 'cn' }));
     const regex = new RegExp(
-      `^<link rel="prefetch" href="${baseHrefCn}/arrow-head-right.min.${hash}.svg" as="image" type="image/svg\\+xml" crossorigin="true">$`
+      `^<link rel="prefetch" href="${baseHrefCn}/arrow-right.min.${hash}.svg" as="image" type="image/svg\\+xml" crossorigin="true">$`
     );
     expect(container.innerHTML).toMatch(regex);
   });
 
   it('should return multiple links', () => {
-    const { container } = render(getIconLinks({ format: 'jsx', icons: ['truck', 'volumeUp', 'mobile'] }));
+    const { container } = render(getIconLinks({ format: 'jsx', icons: ['truck', 'volume-up', 'mobile'] }));
     const regex = new RegExp(
       `^<link rel="prefetch" href="${baseHrefCom}/truck.min.${hash}.svg" as="image" type="image/svg\\+xml" crossorigin="true"><link rel="prefetch" href="${baseHrefCom}/volume-up.min.${hash}.svg" as="image" type="image/svg\\+xml" crossorigin="true"><link rel="prefetch" href="${baseHrefCom}/mobile.min.${hash}.svg" as="image" type="image/svg\\+xml" crossorigin="true">$`
     );
     expect(container.innerHTML).toMatch(regex);
   });
 
-  ICON_NAMES.forEach((iconName: IconNameCamelCase) => {
+  ICON_NAMES.forEach((iconName: IconName) => {
     it(`should match regex for ['${iconName}']`, () => {
       const { container } = render(getIconLinks({ format: 'jsx', icons: [iconName] }));
       const regex = new RegExp(
@@ -104,7 +104,7 @@ describe('format: jsx', () => {
 describe('withoutTags: true', () => {
   it('should return default url', () => {
     const result = getIconLinks({ withoutTags: true });
-    const regex = new RegExp(`^${baseHrefCom}/arrow-head-right.min.${hash}.svg$`);
+    const regex = new RegExp(`^${baseHrefCom}/arrow-right.min.${hash}.svg$`);
 
     expect(result.length).toBe(1);
     expect(result[0]).toMatch(regex);
@@ -112,14 +112,14 @@ describe('withoutTags: true', () => {
 
   it('should return default url for china cdn', () => {
     const result = getIconLinks({ withoutTags: true, cdn: 'cn' });
-    const regex = new RegExp(`^${baseHrefCn}/arrow-head-right.min.${hash}.svg$`);
+    const regex = new RegExp(`^${baseHrefCn}/arrow-right.min.${hash}.svg$`);
 
     expect(result.length).toBe(1);
     expect(result[0]).toMatch(regex);
   });
 
   it('should return multiple urls', () => {
-    const result = getIconLinks({ withoutTags: true, icons: ['truck', 'volumeUp', 'mobile'] });
+    const result = getIconLinks({ withoutTags: true, icons: ['truck', 'volume-up', 'mobile'] });
     const regexTruck = new RegExp(`^${baseHrefCom}/truck.min.${hash}.svg$`);
     const regexVolumeUp = new RegExp(`^${baseHrefCom}/volume-up.min.${hash}.svg$`);
     const regexMobile = new RegExp(`^${baseHrefCom}/mobile.min.${hash}.svg$`);
@@ -130,7 +130,7 @@ describe('withoutTags: true', () => {
     expect(result[2]).toMatch(regexMobile);
   });
 
-  ICON_NAMES.forEach((iconName: IconNameCamelCase) => {
+  ICON_NAMES.forEach((iconName: IconName) => {
     it(`should return icon url for ['${iconName}']`, () => {
       const result = getIconLinks({ withoutTags: true, icons: [iconName] });
       const regex = new RegExp(`^${baseHrefCom}/${paramCase(iconName)}.min.${hash}.svg$`);