+++ packages/components-js/tests/vrt/puppeteer/specs/component-scaling.vrt.ts
@@ -1,19 +1,28 @@
-import { getVisualRegressionStatesTester, vrtTest } from '@porsche-design-system/shared/testing';
+import {
+  getVisualRegressionStatesTester,
+  openPopoversAndHighlightSpacer,
+  vrtTest,
+} from '@porsche-design-system/shared/testing';
 
 const components = [
-  'checkbox-wrapper',
-  'radio-button-wrapper',
-  'link',
   'button',
+  'checkbox-wrapper',
+  'link',
+  'popover',
+  'radio-button-wrapper',
+  'spinner',
   'tag-dismissible',
-  'textarea-wrapper',
   'text-field-wrapper',
+  'textarea-wrapper',
 ];
 
 it.each(components)('should have no visual regression for scaled component %s', async (component) => {
   expect(
     await vrtTest(getVisualRegressionStatesTester(), `${component}-scaling`, `/#${component}`, {
       scenario: async (page) => {
+        if (component === 'popover') {
+          await openPopoversAndHighlightSpacer(page);
+        }
         const client = await page.target().createCDPSession();
         await client.send('Page.enable');
         await client.send('Page.setFontSizes', {