+++ packages/components-react/projects/nextjs/pages/text-field-wrapper.tsx
@@ -522,7 +522,7 @@ const TextFieldWrapperPage: NextPage = (): JSX.Element => {
           description="Multiline description lorem ipsum dolor sit amet, consetetur sadipscing lorem ipsum dolor sit amet"
           state="error"
           message="Multiline message at vero eos et accusam et justo duo dolores et ea rebum."
-          style={{ width: '15rem' }}
+          style={{ maxWidth: '15rem' }}
         >
           <input type="text" defaultValue="Lorem ipsum dolor sit amet, consetetur sadipscing elitr," />
         </PTextFieldWrapper>
@@ -534,7 +534,7 @@ const TextFieldWrapperPage: NextPage = (): JSX.Element => {
           description="Multiline description lorem ipsum dolor sit amet, consetetur sadipscing lorem ipsum dolor sit amet"
           state="error"
           message="Multiline message at vero eos et accusam et justo duo dolores et ea rebum."
-          style={{ width: '240px' }}
+          style={{ maxWidth: '15rem' }}
           theme="dark"
         >
           <input type="text" defaultValue="Lorem ipsum dolor sit amet, consetetur sadipscing elitr," />