+++ packages/components/src/components/checkbox-wrapper/checkbox-wrapper.tsx
@@ -94,7 +94,7 @@ export class CheckboxWrapper {
           <slot />
         </label>
         {hasMessage(this.host, this.message, this.state) && (
-          <StateMessage state={this.state} message={this.message} host={this.host} />
+          <StateMessage state={this.state} message={this.message} theme={this.theme} host={this.host} />
         )}
       </Host>
     );