+++ packages/components/src/components/text-field-wrapper/text-field-wrapper-utils.spec.ts
@@ -1,17 +1,16 @@
+import * as textFieldWrapperUtils from './text-field-wrapper-utils';
 import {
-  getInputPadding,
-  setInputStyles,
-  throwIfUnitLengthExceeded,
-  TextFieldWrapperUnitPosition,
-  hasCounterAndIsTypeText,
-  hasUnitAndIsTypeTextOrNumber,
-  isWithinForm,
-  isType,
   addInputEventListenerForSearch,
   dispatchInputEvent,
+  getInputPaddingLeftOrRight,
+  hasCounterAndIsTypeText,
   hasLocateAction,
+  hasUnitAndIsTypeTextOrNumber,
+  isType,
+  isWithinForm,
+  setInputStyles,
+  throwIfUnitLengthExceeded,
 } from './text-field-wrapper-utils';
-import * as textFieldWrapperUtils from './text-field-wrapper-utils';
 import * as formUtils from '../../utils/form/form-utils';
 import * as getClosestHTMLElementUtils from '../../utils/dom/getClosestHTMLElement';
 
@@ -152,14 +151,12 @@ describe('hasLocateAction()', () => {
   });
 });
 
-describe('getInputPadding()', () => {
-  it.each<[TextFieldWrapperUnitPosition, string]>([
-    ['prefix', '14px 16px 14px 58px'],
-    ['prefix', '14px 16px 14px 58px'],
-    ['suffix', '14px 58px 14px 16px'],
-    ['suffix', '14px 58px 14px 16px'],
-  ])('should for unitPosition: %s return %s', (unitPosition, expected) => {
-    expect(getInputPadding(60, unitPosition)).toBe(expected);
+describe('getInputPaddingLeftOrRight()', () => {
+  it.each<[number, string]>([
+    [60, 'calc(60px - 2px)'],
+    [40, 'calc(40px - 2px)'],
+  ])('should for unitElementWidth: %s return %s', (unitElementWidth, expected) => {
+    expect(getInputPaddingLeftOrRight(unitElementWidth)).toBe(expected);
   });
 });
 
@@ -171,13 +168,22 @@ describe('setInputStyles()', () => {
     expect(input.style.cssText).toBe('');
   });
 
-  it('should set inline padding on input', () => {
+  it('should set inline padding-left var on input', () => {
     const input = getInputElement();
     const unitElement = getCounterElement();
     Object.defineProperty(unitElement, 'offsetWidth', { value: 60 });
     setInputStyles(input, unitElement, 'prefix');
 
-    expect(input.style.cssText).toBe('padding: 14px 16px 14px 58px !important;');
+    expect(input.style.cssText).toBe('--p-internal-text-field-input-padding-left: calc(60px - 2px) !important;');
+  });
+
+  it('should set inline padding-right var on input', () => {
+    const input = getInputElement();
+    const unitElement = getCounterElement();
+    Object.defineProperty(unitElement, 'offsetWidth', { value: 60 });
+    setInputStyles(input, unitElement, 'suffix');
+
+    expect(input.style.cssText).toBe('--p-internal-text-field-input-padding-right: calc(60px - 2px) !important;');
   });
 });