+++ packages/components/src/components/text-field-wrapper/text-field-wrapper-utils.ts
@@ -1,6 +1,7 @@
 import { getClosestHTMLElement, hasCounter } from '../../utils';
 import type { IconName } from '../../types';
-import { spacingStaticMedium } from '@porsche-design-system/utilities-v2';
+import { borderWidthBase } from '@porsche-design-system/utilities-v2';
+import { cssVariableInputPaddingLeft, cssVariableInputPaddingRight } from './text-field-wrapper-styles';
 
 export const UNIT_POSITIONS = ['prefix', 'suffix'] as const;
 export type TextFieldWrapperUnitPosition = typeof UNIT_POSITIONS[number];
@@ -14,11 +15,8 @@ export const isType = (inputType: string, typeToValidate: string): boolean => in
 export const isWithinForm = (host: HTMLElement): boolean => !!getClosestHTMLElement(host, 'form');
 export const hasLocateAction = (icon: IconName): boolean => icon === 'locate';
 
-export const getInputPadding = (unitElementWidth: number, unitPosition: TextFieldWrapperUnitPosition): string => {
-  const topBottomPadding = '14px';
-  return unitPosition === 'prefix'
-    ? `${topBottomPadding} ${spacingStaticMedium} ${topBottomPadding} ${unitElementWidth - 2}px` // -2px to compensate border
-    : `${topBottomPadding} ${unitElementWidth - 2}px ${topBottomPadding} ${spacingStaticMedium}`; // -2px to compensate border
+export const getInputPaddingLeftOrRight = (unitElementWidth: number): string => {
+  return `calc(${unitElementWidth}px - ${borderWidthBase})`;
 };
 
 export const setInputStyles = (
@@ -27,7 +25,14 @@ export const setInputStyles = (
   unitPosition: TextFieldWrapperUnitPosition
 ): void => {
   if (unitOrCounterElement) {
-    input.style.setProperty('padding', getInputPadding(unitOrCounterElement.offsetWidth, unitPosition), 'important');
+    input.style.removeProperty(cssVariableInputPaddingLeft);
+    input.style.removeProperty(cssVariableInputPaddingRight);
+
+    input.style.setProperty(
+      unitPosition === 'prefix' ? cssVariableInputPaddingLeft : cssVariableInputPaddingRight,
+      getInputPaddingLeftOrRight(unitOrCounterElement.offsetWidth),
+      'important'
+    );
   }
 };